refactor: move uppyService and useUpload composable to web-pkg
This is needed since we should (and soon must) no import from web-runtime.
