+++ packages/web-app-files/src/HandleUpload.ts
@@ -9,8 +9,7 @@ import { Store } from 'vuex'
 import { RouteLocationNormalizedLoaded } from 'vue-router'
 import { Resource, SpaceResource } from 'web-client/src'
 import { urlJoin } from 'web-client/src/utils'
-import { UppyResource } from 'web-runtime/src/composables/upload'
-import { UppyService } from 'web-runtime/src/services/uppyService'
+import { UppyService, UppyResource } from 'web-pkg/src/services/uppy'
 import { ResourceConflict } from './helpers/resource'
 import { locationPublicLink } from 'web-pkg/src/router/public'
 import { locationSpacesGeneric } from 'web-pkg/src/router/spaces'
