+++ packages/web-app-files/tests/unit/HandleUpload.spec.ts
@@ -1,9 +1,8 @@
 import Uppy, { UppyFile, State, UIPlugin } from '@uppy/core'
 import { HandleUpload } from '../../src/HandleUpload'
 import { mock, mockDeep } from 'jest-mock-extended'
-import { UppyResource } from 'web-runtime/src/composables/upload'
 import { Resource, SpaceResource } from 'web-client/src'
-import { UppyService } from 'web-runtime/src/services/uppyService'
+import { UppyService, UppyResource } from 'web-pkg/src/services/uppy'
 import { RouteLocationNormalizedLoaded } from 'vue-router'
 import { ref, unref } from 'vue'
 import { ClientService } from 'web-pkg/src'
