+++ packages/web-pkg/package.json
@@ -17,6 +17,11 @@
     "@casl/ability": "^6.3.3",
     "@casl/vue": "^2.2.1",
     "@microsoft/fetch-event-source": "^2.0.1",
+    "@uppy/core": "^3.2.0",
+    "@uppy/drop-target": "^2.0.0",
+    "@uppy/tus": "^3.1.0",
+    "@uppy/utils": "^5.3.0",
+    "@uppy/xhr-upload": "^3.0.1",
     "@vueuse/core": "^10.0.0",
     "axios": "^0.27.2 || ^1.0.0",
     "design-system": "workspace:@ownclouders/design-system@*",
