+++ packages/web-pkg/src/composables/index.ts
@@ -22,3 +22,4 @@ export * from './viewMode'
 export * from './search'
 export * from './sse'
 export * from './passwordPolicyService'
+export * from './upload'
