+++ packages/web-pkg/src/composables/upload/useUpload.ts
@@ -8,50 +8,9 @@ import {
   useStore
 } from 'web-pkg/src/composables'
 import { computed, unref, watch } from 'vue'
-import { UppyService } from '../../services/uppyService'
+import { UppyService } from '../../services/uppy/uppyService'
 import { v4 as uuidV4 } from 'uuid'
 import { useGettext } from 'vue3-gettext'
-
-export interface UppyResource {
-  id?: string
-  source: string
-  name: string
-  isFolder: boolean
-  type: string
-  size: number
-  data: Blob
-  isRemote: boolean
-  meta: {
-    // IMPORTANT: must only contain primitive types, complex types won't be serialized properly!
-    name?: string
-    mtime?: number
-    // current space & folder
-    spaceId: string | number
-    spaceName: string
-    driveAlias: string
-    driveType: string
-    currentFolder: string // current folder path during upload initiation
-    currentFolderId?: string | number
-    fileId?: string | number
-    // upload data
-    uppyId?: string
-    relativeFolder: string
-    relativePath: string
-    tusEndpoint: string
-    uploadId: string
-    topLevelFolderId?: string
-    // route data
-    routeName?: string
-    routeDriveAliasAndItem?: string
-    routeShareId?: string
-  }
-  tus?: {
-    endpoint: string
-  }
-  xhrUpload?: {
-    endpoint: string
-  }
-}
 
 interface UploadOptions {
   uppyService: UppyService
