+++ packages/web-pkg/src/services/uppy/index.ts
@@ -1,0 +1,2 @@
+export * from './types'
+export * from './uppyService'
