+++ packages/web-runtime/package.json
@@ -11,10 +11,7 @@
     "@popperjs/core": "^2.11.5",
     "@sentry/vue": "7.55.2",
     "@uppy/core": "^3.2.0",
-    "@uppy/drop-target": "^2.0.0",
-    "@uppy/tus": "^3.1.0",
     "@uppy/utils": "^5.3.0",
-    "@uppy/xhr-upload": "^3.0.1",
     "@vueuse/head": "1.1.26",
     "axios": "^1.0.0",
     "design-system": "workspace:@ownclouders/design-system@*",
